+++ ios/Event+JSON.swift
@@ -417,11 +417,15 @@ extension PlaybackSpeedChangedEvent: JsonConvertible {
 extension CueEnterEvent: JsonConvertible {
     func toJSON() -> [AnyHashable: Any] {
         toEventJSON {
-            [
+            var json: [AnyHashable: Any] = [
                 "start": startTime,
                 "end": endTime,
-                "text": text
+                "text": text,
             ]
+            if let imagePngData = image?.pngData() {
+                json["imageUri"] = "data:image/png;base64,\(imagePngData.base64EncodedString())"
+            }
+            return json
         }
     }
 }